requirements.txt
streamlit
supabase
python-dotenv
pandas
google-genai
google-adk
pydantic
tenacity
